+++ src/components/helpers/Select.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+function Select() {
+    return (
+        <div>
+            <select class="js-example-basic-single" name="state">
+                <option value="AL">Alabama</option>
+                ...
+                <option value="WY">Wyoming</option>
+            </select>
+        </div>
+    );
+}
+
+export default Select;
